verification before deleting allusers so no error appear. added password to the entity

+++ app/src/main/java/com/example/android_studio_project/data/room/vm/UserViewModel.kt
@@ -40,7 +40,10 @@ class UserViewModel(application: Application) : AndroidViewModel(application) {
 
     fun deleteAllUsers() {
         viewModelScope.launch(Dispatchers.IO) {
-            repository.deleteAllUsers()
+            val users = readAllUsers.value
+            if (!users.isNullOrEmpty()) {
+                repository.deleteAllUsers()
+            }
         }
     }
 }
